modification of the Lesson model and creation of api routes for the Lesson imge upload

+++ next-app/src/hooks/mutations/use-update-user-avatar.ts
@@ -15,10 +15,11 @@ export const useUpdateUserAvatar = () => {
   return useMutation({
     mutationFn: async (formData: FormData) => {
       const img = data?.user?.image?.split("/").pop();
-
-      await fetch(`http://localhost:8000/avatar/${img}`, {
-        method: "DELETE",
-      });
+      if (img) {
+        await fetch(`http://localhost:8000/avatar/${img}`, {
+          method: "DELETE",
+        });
+      }
 
       const res = await fetch(`http://localhost:8000/avatar`, {
         method: "POST",
